Refactoring
Refactored a lot of code to be better structured and use a little less memory.
Some notible restructurings are the change to factory manager to contain a list of unique factory registries and a count of each. And the change to the saving and loading system moved into a state and a state preserver interface for more dynamic handling of saving.